what

+++ Assignment3/barber.c
@@ -61,7 +61,7 @@ def Customer():
 #define LOWER_MASK 0x7fffffffUL
 
 /* This calls assembler instruction rdrand and places the random number value into *arand */
-int rdrand64_step (uint64_t *arand)
+int rdrand64_step (unsigned long long *arand)
 {
         unsigned char ok;
         __asm__ __volatile__ ("rdrand %0; setc %1" : "=r" (*arand), "=qm" (ok));
